Add app.rs and ui.rs to tui_util
