+++ src/parts.rs
@@ -2,26 +2,28 @@
 use std::fs;
 
 
-pub struct Parts {
+pub struct PartsManager {
     search_paths: Vec<String>,
-    part_names: Vec<String>,
+    parts: Vec<String>,
+    mounted_parts: Vec<String>,
 }
 
-impl Parts {
+impl PartsManager {
     // create instance of part managaer
-    pub fn new() -> Parts {
+    pub fn new() -> PartsManager {
         // get directories of parts
-        let mut parts = Vec::new();
+        let mut p = Vec::new();
         match fs::read_dir("parts/") {
             Ok(paths) => for path in paths {
-                parts.push(path.unwrap().path().to_str().unwrap().to_owned())
+                p.push(path.unwrap().path().to_str().unwrap().to_owned())
             },
             Err(why) => println!("! {:?}", why.kind()),
         }
 
-        Parts { 
+        PartsManager { 
             search_paths: vec!["parts/".to_owned()],
-            part_names: parts,
+            parts: p,
+            mounted_parts: Vec::new(),
         }
     }
 
@@ -32,17 +34,28 @@ impl Parts {
 
     // returns part diectories
     pub fn get_part_names(&self) -> Vec<String> {
-        self.part_names.clone()
+        self.parts.clone()
     }
 
     // returns only lua parts
-    pub fn get_lua_parts(&self) -> Vec<String> {
+    pub fn get_mounted_lua_parts(&self) -> Vec<String> {
         let mut luas = Vec::new();
-        for part in self.part_names.clone() {
+        for part in self.mounted_parts.clone() {
             if part.contains(".lua") {
                 luas.push(part)
             }
         }
+
         luas
+    }
+
+    // mounts part to core
+    pub fn mount(&mut self, part: &str) {
+        if self.parts.contains(&part.to_owned()) {
+            self.mounted_parts.push(part.to_owned());
+            println!("Successfully mounted {} to core", part);
+        } else {
+            println!("Failed to mount. Reason: Unknown part name \"{}\"", part);
+        }
     }
 }
