+++ src/tui_util/mod.rs
@@ -1,1 +1,3 @@
 pub mod backend;
+pub mod app;
+pub mod ui;
